Move the code from capslock into internal package with own license.
